+++ rbx_dom_weak/src/lib.rs
@@ -70,11 +70,12 @@
 
 #![warn(rust_2018_idioms)]
 
-mod unresolved_value;
 mod brick_color;
 mod id;
 mod instance;
+mod shared_string;
 mod tree;
+mod unresolved_value;
 mod value;
 
 pub use crate::{
@@ -83,5 +84,6 @@ pub use crate::{
     id::RbxId,
     instance::{RbxInstanceProperties, RbxInstance},
     tree::{RbxTree, Descendants},
+    shared_string::SharedString,
     value::*,
 };