+++ rbx_xml/tests/round_trip.rs
@@ -13,6 +13,7 @@ static TEST_MODELS: &[&str] = &[
     include_str!("../test-files/parts.rbxmx"),
     include_str!("../test-files/sound.rbxmx"),
     include_str!("../test-files/terrain.rbxmx"),
+    include_str!("../test-files/union.rbxmx"),
 ];
 
 #[test]